requirements.txt
aiogram
python-dotenv
googletrans==3.1.0a0
emoji-country-flag